+++ package.json
@@ -9,9 +9,12 @@
   },
   "dependencies": {
     "audio-loader": "^0.5.0",
+    "fast-xml-parser": "^4.0.8",
     "midi-file": "^1.2.3",
+    "musicxml-iterator": "^0.0.0",
     "sample-player": "^0.5.5",
-    "vue": "^3.2.25"
+    "vue": "^3.2.25",
+    "webmscore": "^0.21.0"
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "^2.2.0",
