+++ src/App.vue
@@ -13,7 +13,10 @@
  * <list
  */
 import Mixer from './components/mixer.vue';
-import { MIDIFile } from './components/midi';
+// import { MIDIFile } from './components/midi';
+// import iterator from 'musicxml-iterator';
+// import WebMscore from 'webmscore';
+import {XMLParser} from 'fast-xml-parser';
 export default {
   components:{
     Mixer
@@ -76,6 +79,11 @@ export default {
     }
   },
   async mounted(){
+    // const r = await fetch('Chattanooga_Choo_Choo.musicxml');
+    // const b = await r.text();
+    // const score = new XMLParser().parse(b);
+    // console.log(score);
+    // return
     this.readScores();
     if (window.location.search){
       const q = window.location.search.slice(1).split('&').map(s=>s.split('=')).reduce((o,d)=>Object.assign(o,{[d[0]]:d[1]}),{});
